🚚 api: move ods utils to enrolment
