+++ api/src/certification/enrolment/domain/services/certification-candidates-ods-service.js
@@ -4,10 +4,10 @@ import _ from 'lodash';
 import { CERTIFICATION_CANDIDATES_ERRORS } from '../../../../../lib/domain/constants/certification-candidates-errors.js';
 import { CertificationCandidatesError } from '../../../../../lib/domain/errors.js';
 import { CertificationCandidate } from '../../../../../lib/domain/models/CertificationCandidate.js';
-import * as readOdsUtils from '../../../../shared/infrastructure/utils/ods/read-ods-utils.js';
 import * as mailCheckImplementation from '../../../../shared/mail/infrastructure/services/mail-check.js';
 import { ComplementaryCertificationKeys } from '../../../shared/domain/models/ComplementaryCertificationKeys.js';
 import { getTransformationStructsForPixCertifCandidatesImport } from '../../infrastructure/files/candidates-import/candidates-import-transformation-structures.js';
+import * as readOdsUtils from '../../infrastructure/utils/ods/read-ods-utils.js';
 import { Subscription } from '../models/Subscription.js';
 
 export { extractCertificationCandidatesFromCandidatesImportSheet };
