+++ api/src/certification/enrolment/infrastructure/files/candidates-import/fill-candidates-import-sheet.js
@@ -3,8 +3,8 @@ import * as url from 'node:url';
 import _ from 'lodash';
 
 import { CertificationCandidate } from '../../../../../../lib/domain/models/index.js';
-import * as readOdsUtils from '../../../../../shared/infrastructure/utils/ods/read-ods-utils.js';
-import * as writeOdsUtils from '../../../../../shared/infrastructure/utils/ods/write-ods-utils.js';
+import * as readOdsUtils from '../../utils/ods/read-ods-utils.js';
+import * as writeOdsUtils from '../../utils/ods/write-ods-utils.js';
 import { CandidateData } from './CandidateData.js';
 import {
   EXTRA_EMPTY_CANDIDATE_ROWS,
