+++ api/src/certification/enrolment/infrastructure/utils/ods/added-cell-option.js
