+++ api/tests/certification/enrolment/integration/infrastructure/ods/read-ods-utils_test.js
@@ -6,20 +6,20 @@ const { readFile } = promises;
 
 import _ from 'lodash';
 
-import { getI18n } from '../../../../tooling/i18n/i18n.js';
+import { getI18n } from '../../../../../tooling/i18n/i18n.js';
 const i18n = getI18n();
 
 import * as url from 'node:url';
 
-import { UnprocessableEntityError } from '../../../../../lib/application/http-errors.js';
-import { getTransformationStructsForPixCertifCandidatesImport } from '../../../../../lib/infrastructure/files/candidates-import/candidates-import-transformation-structures.js';
+import { UnprocessableEntityError } from '../../../../../../lib/application/http-errors.js';
+import { getTransformationStructsForPixCertifCandidatesImport } from '../../../../../../src/certification/enrolment/infrastructure/files/candidates-import/candidates-import-transformation-structures.js';
 import {
   extractTableDataFromOdsFile,
   getContentXml,
   getSheetDataRowsFromOdsBuffer,
   validateOdsHeaders,
-} from '../../../../../src/shared/infrastructure/utils/ods/read-ods-utils.js';
-import { catchErr, expect } from '../../../../test-helper.js';
+} from '../../../../../../src/certification/enrolment/infrastructure/utils/ods/read-ods-utils.js';
+import { catchErr, expect } from '../../../../../test-helper.js';
 const __dirname = url.fileURLToPath(new URL('.', import.meta.url));
 
 describe('Integration | Shared | Infrastructure | Utils | Ods | read-ods-utils', function () {
