+++ api/tests/certification/enrolment/integration/infrastructure/ods/write-ods-utils_test.js
@@ -6,8 +6,8 @@ const { writeFile, unlink } = promises;
 
 import * as url from 'node:url';
 
-import { AddedCellOption } from '../../../../../src/shared/infrastructure/utils/ods/added-cell-option.js';
-import { getContentXml } from '../../../../../src/shared/infrastructure/utils/ods/read-ods-utils.js';
+import { AddedCellOption } from '../../../../../../src/certification/enrolment/infrastructure/utils/ods/added-cell-option.js';
+import { getContentXml } from '../../../../../../src/certification/enrolment/infrastructure/utils/ods/read-ods-utils.js';
 import {
   addCellToEndOfLineWithStyleOfCellLabelled,
   addTooltipOnCell,
@@ -16,8 +16,8 @@ import {
   makeUpdatedOdsByContentXml,
   updateXmlRows,
   updateXmlSparseValues,
-} from '../../../../../src/shared/infrastructure/utils/ods/write-ods-utils.js';
-import { expect } from '../../../../test-helper.js';
+} from '../../../../../../src/certification/enrolment/infrastructure/utils/ods/write-ods-utils.js';
+import { expect } from '../../../../../test-helper.js';
 const __dirname = url.fileURLToPath(new URL('.', import.meta.url));
 
 describe('Integration | Shared | Infrastructure | Utils | Ods | write-ods-utils', function () {
